+++ zoom.cpp
@@ -2,7 +2,7 @@
 #include "zoom.h"
 using namespace std;
 
-Zoom::Zoom(int x, int y, double scale, double percentage) : x(x), y(y), scale(scale), percentage(percentage)
+Zoom::Zoom(long double x, long double y, double scale, double percentage) : x(x), y(y), scale(scale), percentage(percentage)
 {
 }
 Zoom::~Zoom()
